+++ cstar/base/__init__.py
@@ -1,5 +1,5 @@
 from cstar.base.component import Component
-from cstar.base.basemodel import BaseModel
+from cstar.base.base_model import BaseModel
 from cstar.base.additional_code import AdditionalCode
 from cstar.base.input_dataset import InputDataset
