+++ cstar/marbl/__init__.py
@@ -1,4 +1,4 @@
-from cstar.marbl.basemodel import MARBLBaseModel
+from cstar.marbl.base_model import MARBLBaseModel
 from cstar.marbl.component import MARBLComponent
 
 __all__ = [
